delete old implementation of canonical
